pyproject.toml
[tool.poetry]
name = "datapipe-image-moderation"
version = "0.1.0"
description = ""
authors = ["Sultan Zakariya <sultan@epoch8.co>"]

exclude = ["tests"]

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
requests = "^2.31.0"
google-cloud-vision = "^3.4.4"
yandexcloud = "^0.224.0"
datapipe-core = {version=">=0.13.0-alpha.5, <0.14", allow-prereleases = true}
pandas = "^1.2.0"
numpy = "^1.21.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pylint = "^2.17.4"
pylint-pydantic = "^0.1.8"
mypy = "^1.3.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
